backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.0
fastapi==0.110.1
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
mccabe==0.7.0
passlib==1.7.4
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
